Fix memory leak (#47)
* a

* vulnerability?

* removeundici
