+++ src/events/tagDetect.ts
@@ -1,10 +1,9 @@
 import { eventModule, EventType } from "@sern/handler";
 import { ActionRowBuilder, ButtonBuilder, ButtonStyle, EmbedBuilder, Message } from "discord.js";
-import { createRequire } from "module";
 import { FuzzyMatcher } from "../utils/FuzzyMatcher.js";
 import type { TagData, TagMessage } from "typings";
 import { TagList } from "#constants";
-const require = createRequire(import.meta.url);
+import { require } from '#utils'
 const file: TagData[] = require(TagList);
 
 export default eventModule({
