--- src/utils/require.ts
@@ -1,2 +1,0 @@
-import { createRequire } from "module";
-export const require = createRequire(import.meta.url);
